Development

+++ src/main/resources/db/migrations/V5__create_room_schema.sql
@@ -1,5 +1,5 @@
 CREATE TABLE rooms (
-    room_id INT PRIMARY KEY AUTO_INCREMENT,
+    room_id BIGINT PRIMARY KEY AUTO_INCREMENT,
     name VARCHAR(255),
     capacity INT NOT NULL,
     location VARCHAR(100) NOT NULL
